chore: address PR comments

+++ modules/engines/engine-javascript-graalium/src/main/java/org/eclipse/dirigible/engine/js/graalium/execution/modules/DirigibleCoreModuleResolver.java
@@ -26,9 +26,11 @@ public class DirigibleCoreModuleResolver {
     private static final Pattern DIRIGIBLE_CORE_MODULE_SIGNATURE_PATTERN = Pattern.compile("(@dirigible)(\\/)(\\w+)"); // e.g. @dirigible/core  => $1=dirigible $2=/ $3=core
 
     private final DirigibleCoreModuleESMProxyGenerator dirigibleCoreModuleESMProxyGenerator;
+    private final Path cacheDirectoryPath;
 
-    public DirigibleCoreModuleResolver() {
+    public DirigibleCoreModuleResolver(Path cacheDirectoryPath) {
         dirigibleCoreModuleESMProxyGenerator = new DirigibleCoreModuleESMProxyGenerator();
+        this.cacheDirectoryPath = cacheDirectoryPath;
     }
 
     public boolean isCoreModule(String moduleToResolve) {
@@ -42,9 +44,15 @@ public class DirigibleCoreModuleResolver {
         }
 
         String coreModuleName = modulePathMatcher.group(3);
-        String coreModuleContent = dirigibleCoreModuleESMProxyGenerator.generate(coreModuleName, "");
 
-        File coreModuleGeneratedFile = new File("/Users/c5326377/work/dirigible/core-modules/" + coreModuleName + ".mjs");
+        Path coreModuleGeneratedPath = cacheDirectoryPath.resolve(coreModuleName + ".mjs");
+        File coreModuleGeneratedFile = coreModuleGeneratedPath.toFile();
+
+        if (coreModuleGeneratedFile.exists()) {
+            return coreModuleGeneratedFile.toPath();
+        }
+
+        String coreModuleContent = dirigibleCoreModuleESMProxyGenerator.generate(coreModuleName, "");
         try {
             FileUtils.write(coreModuleGeneratedFile, coreModuleContent, StandardCharsets.UTF_8);
         } catch (IOException e) {
